safari-600.7-branch/LayoutTests/fast/dom/document-set-body-expected.txt

Tests setting document.body

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document1.body = iframe1 threw exception Error: HierarchyRequestError: DOM Exception 3.
PASS iframe1.parentNode is document.body
PASS document1.body = document1.createElement('iframe') threw exception Error: HierarchyRequestError: DOM Exception 3.
PASS document1.body != document.body is true
PASS document1.body is body1
PASS successfullyParsed is true

TEST COMPLETE
